requirements.txt
openpyxl~=3.0.9
Pillow~=8.3.1
PySimpleGUI~=4.40
